+++ src/SqlArtisan/Internal/SqlPart/Keywords.cs
@@ -55,6 +55,7 @@ internal static class Keywords
     internal const string Full = "FULL";
     internal const string Greatest = "GREATEST";
     internal const string Group = "GROUP";
+    internal const string GroupConcat = "GROUP_CONCAT";
     internal const string Having = "HAVING";
     internal const string In = "IN";
     internal const string Inner = "INNER";
@@ -75,6 +76,7 @@ internal static class Keywords
     internal const string Lengthb = "LENGTHB";
     internal const string Like = "LIKE";
     internal const string Limit = "LIMIT";
+    internal const string Listagg = "LISTAGG";
     internal const string Locked = "LOCKED";
     internal const string Lower = "LOWER";
     internal const string Lpad = "LPAD";
@@ -126,10 +128,12 @@ internal static class Keywords
     internal const string Rpad = "RPAD";
     internal const string Rtrim = "RTRIM";
     internal const string Select = "SELECT";
+    internal const string Separator = "SEPARATOR";
     internal const string Set = "SET";
     internal const string Sign = "SIGN";
     internal const string Skip = "SKIP";
     internal const string Sqrt = "SQRT";
+    internal const string StringAgg = "STRING_AGG";
     internal const string Substr = "SUBSTR";
     internal const string Substrb = "SUBSTRB";
     internal const string Sum = "SUM";